added backend seperate backend to store and retrive campaigns user dontated to, UI updates, remove duplicate code
