+++ pages/campaigns/hot.js
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
-import { Header, Card, Icon } from 'semantic-ui-react';
+import { Segment, Header, Button, Card, Icon } from 'semantic-ui-react';
 import factory from '../../ethereum/factory';
 import Layout from '../../components/Layout.js';
 import { Link } from '../../routes';
 import Campaign from '../../ethereum/campaign';
 import web3 from '../../ethereum/web3';
+import usefulMethods from '../../usefulMethods';
 
 class Hot extends Component {
   static async getInitialProps() {
@@ -29,16 +30,6 @@ class Hot extends Component {
     return { campaigns, campaignsArray };
   }
 
-  countdown(start, target) {
-    const startWithMilli = parseInt(start) * 1000
-    const targetWithMilli = parseInt(target) * 1000
-    const diff = startWithMilli + targetWithMilli - Date.now()
-    const hoursLeft = diff / 60 / 60 / 1000
-    let output;
-    hoursLeft < 0 ? output = 0 :  output = hoursLeft.toFixed(1)
-    return output;
-  }
-
   renderCampaigns() {
     let editorPicks = [];
     const allItems = this.props.campaignsArray.map((data, index) => {
@@ -46,7 +37,7 @@ class Hot extends Component {
         key: index,
         image: `https://picsum.photos/300/200?image=${ Math.ceil(Math.random() * 100)}`,
         header: data.description,
-        meta: `${this.countdown(data.start, data.timeGoal)} hours remaining`,
+        meta: `${usefulMethods.countdown(data.start, data.timeGoal)} hours remaining`,
         description: (
           <Link route={`/campaigns/${this.props.campaigns[index]}`}>
             <a>View Campaign</a>
@@ -61,14 +52,16 @@ class Hot extends Component {
     for(let i = 0; i < 3; i++) {
       editorPicks.push(activeItems[Math.floor(Math.random()*activeItems.length)]);
     }
-    return <Card.Group items={editorPicks} />;
+    return <Card.Group items={editorPicks} centered />;
   }
 
   render() {
     return (
       <Layout>
         <Header as='h1' textAlign="center">Hot Campaigns</Header>
-        {this.renderCampaigns()}
+        <Segment>
+          {this.renderCampaigns()}
+        </Segment>
       </Layout>
     )
   }
